refactor(testmessage): refactor testmessage imports
